+++ tests/testthat/test-select.axes.R
@@ -1,0 +1,14 @@
+## Test
+test_that("select.axes works", {
+
+    ## Sanitizing
+    # expect_error(select.axes())
+
+    # ## Right output
+    # expect_is(
+    #     select.axes()
+    #     , "class")
+    # expect_equal(
+    #     dim(select.axes())
+    #     , dim)
+})